View: Statistiques labels and display

+++ Projet2CP/Model/IPromoStatistics.vb
@@ -1,7 +1,7 @@
 ﻿Public Interface IPromoStatistics
     Inherits IStatistics
 
-    Function getEtudiantDistribution() As List(Of Double)
+    Function getEtudiantDistribution() As List(Of Integer)
 
     Function getTauxReussite() As Object
